feat(tt): support sections in tutorial index and solution
