+++ apps/total-typescript/src/pages/tutorials/[module]/[section]/[lesson]/solution.tsx
@@ -1,8 +1,9 @@
 import React from 'react'
 import ExerciseTemplate from 'templates/exercise-template'
 import {GetStaticPaths, GetStaticProps} from 'next'
+import {Lesson} from '@skillrecordings/skill-lesson/schemas/lesson'
 import {getAllTutorials, getTutorial} from 'lib/tutorials'
-import {getExercise} from 'lib/exercises'
+import {getExercise, Exercise} from 'lib/exercises'
 import {VideoResourceProvider} from '@skillrecordings/skill-lesson/hooks/use-video-resource'
 import {LessonProvider} from '@skillrecordings/skill-lesson/hooks/use-lesson'
 import {ModuleProgressProvider} from 'video/module-progress'
@@ -16,10 +17,10 @@ export const getStaticProps: GetStaticProps = async (context) => {
 
   return {
     props: {
-      exercise,
+      solution: exercise.solution,
       module,
-      transcript: exercise.transcript,
-      videoResourceId: exercise.videoResourceId,
+      transcript: exercise.solution?.transcript,
+      videoResourceId: exercise.solution?.videoResourceId,
     },
     revalidate: 10,
   }
@@ -28,32 +29,36 @@ export const getStaticProps: GetStaticProps = async (context) => {
 export const getStaticPaths: GetStaticPaths = async (context) => {
   const tutorials = await getAllTutorials()
 
-  const paths = tutorials.reduce((acc: any[], tutorial: any) => {
-    return [
-      ...acc,
-      ...tutorial.lessons.map((exercise: any) => {
-        return {
-          params: {
-            module: tutorial.slug.current,
-            lesson: exercise.slug,
-          },
-        }
-      }),
-    ]
-  }, [])
+  const paths = tutorials.flatMap((tutorial: any) => {
+    return (
+      tutorial.sections?.flatMap((section: any) => {
+        return (
+          section.lessons
+            ?.filter(({_type}: Lesson) => _type === 'exercise')
+            .map((exercise: Exercise) => ({
+              params: {
+                module: tutorial.slug.current,
+                section: section.slug,
+                exercise: exercise.slug,
+              },
+            })) || []
+        )
+      }) || []
+    )
+  })
 
   return {paths, fallback: 'blocking'}
 }
 
-const ExercisePage: React.FC<any> = ({
-  exercise,
+const ExerciseSolution: React.FC<any> = ({
+  solution,
   module,
   transcript,
   videoResourceId,
 }) => {
   return (
     <ModuleProgressProvider moduleSlug={module.slug.current}>
-      <LessonProvider lesson={exercise} module={module}>
+      <LessonProvider lesson={solution} module={module}>
         <VideoResourceProvider videoResourceId={videoResourceId}>
           <ExerciseTemplate transcript={transcript} />
         </VideoResourceProvider>
@@ -62,4 +67,4 @@ const ExercisePage: React.FC<any> = ({
   )
 }
 
-export default ExercisePage
+export default ExerciseSolution
